Render skill references in server-side instructions HTML
The server markdown-to-HTML converter (convertMarkdownToBlockHtml) handled
knowledge and tool tags but not skill references, so <skill .../> was
HTML-escaped into literal text in instructionsHtml. The builder loads that
HTML, so edits made via the skill_authoring tool rendered as raw XML.

Wire skill references through the same path as knowledge and tool:
- Split SkillNode into a schema-only base (skill_builder/SkillNode) and the
  React view (input_bar/SkillNode), mirroring ToolNode/ToolNodeWithView, so
  the server renderer can register it without pulling in React.
- Register the base node in the server extension list and recover standalone
  <skill>/<unavailable_skill> lines.
- Preserve the id attribute on skill/unavailable_skill during the strip step.

Two rendering bugs fixed along the way:
- skillUnavailable was parse-only but had no renderHTML, so the static
  renderer emitted a malformed skillunavailable="false/" attribute (the
  client only avoided it because DOMPurify drops all but id/name/icon). Make
  it render nothing; the unavailable state is encoded by the tag name.
- A childless 2-tuple renderHTML self-closes as <skill .../> in the static
  string renderer, and HTML5 parsers ignore self-close on unknown tags, so
  trailing text got swallowed into the node. Render an explicit chip child
  like ToolNode/KnowledgeNode to force a properly closed element.

Add a targeted backfill for skills whose stored HTML escaped their skill
references.
