+++ front/components/editor/extensions/input_bar/SkillNode.tsx
@@ -1,175 +1,18 @@
-import { SkillNodeComponent } from "@app/components/editor/input_bar/SkillNodeComponent";
 import {
-  parseSkillReferenceTag,
-  SKILL_REFERENCE_TAG_REGEX_BEGINNING,
-  SKILL_TAG_NAME,
-  serializeSkillTag,
-  serializeUnavailableSkillTag,
-  UNAVAILABLE_SKILL_TAG_NAME,
-} from "@app/lib/skills/format";
-import { isString } from "@app/types/shared/utils/general";
-import { Node } from "@tiptap/core";
+  SKILL_NODE_TYPE,
+  SkillNode as SkillNodeBase,
+} from "@app/components/editor/extensions/skill_builder/SkillNode";
+import { SkillNodeComponent } from "@app/components/editor/input_bar/SkillNodeComponent";
 import { ReactNodeViewRenderer } from "@tiptap/react";
 
-export type SkillNodeAttributes = {
-  skillId: string;
-  skillIcon?: string | null;
-  skillName: string;
-  skillUnavailable?: boolean;
-};
+export type { SkillNodeAttributes } from "@app/components/editor/extensions/skill_builder/SkillNode";
+export { SKILL_NODE_TYPE };
 
-export const SKILL_NODE_TYPE = "skill";
-
-declare module "@tiptap/core" {
-  interface Commands<ReturnType> {
-    skillNode: {
-      insertSkillNode: (attrs: SkillNodeAttributes) => ReturnType;
-    };
-  }
-}
-
-// TODO(2026-05-02 aubin): Check whether we can share logic with KnowledgeNode,
-// for example through a base extension.
-export const SkillNode = Node.create({
-  name: SKILL_NODE_TYPE,
-  group: "inline",
-  inline: true,
-  atom: true,
-  selectable: false,
-
-  addAttributes() {
-    return {
-      skillId: {
-        default: null,
-        parseHTML: (element) => element.getAttribute("id"),
-        renderHTML: (attributes) =>
-          isString(attributes.skillId) ? { id: attributes.skillId } : {},
-      },
-      skillName: {
-        default: null,
-        parseHTML: (element) => element.getAttribute("name"),
-        renderHTML: (attributes) =>
-          isString(attributes.skillName) ? { name: attributes.skillName } : {},
-      },
-      skillIcon: {
-        default: null,
-        parseHTML: (element) => element.getAttribute("icon"),
-        renderHTML: (attributes) =>
-          isString(attributes.skillIcon) ? { icon: attributes.skillIcon } : {},
-      },
-      skillUnavailable: {
-        default: false,
-        parseHTML: (element) =>
-          element.tagName.toLowerCase() === UNAVAILABLE_SKILL_TAG_NAME,
-      },
-    };
-  },
-
-  // HTML serialization and deserialization.
-  parseHTML() {
-    return [{ tag: SKILL_TAG_NAME }, { tag: UNAVAILABLE_SKILL_TAG_NAME }];
-  },
-
-  renderHTML({ node, HTMLAttributes }) {
-    if (node.attrs.skillUnavailable === true) {
-      return [
-        UNAVAILABLE_SKILL_TAG_NAME,
-        isString(node.attrs.skillId) ? { id: node.attrs.skillId } : {},
-      ];
-    }
-
-    return [SKILL_TAG_NAME, HTMLAttributes];
-  },
-
-  renderText({ node }) {
-    if (node.attrs.skillUnavailable === true) {
-      return "/Unavailable skill";
-    }
-
-    return `/${node.attrs.skillName ?? "skill"}`;
-  },
-
+// Interactive variant of SkillNode that adds the React node view. The
+// schema-only base lives in skill_builder/SkillNode so server-side code can
+// register the node without pulling in React.
+export const SkillNode = SkillNodeBase.extend({
   addNodeView() {
     return ReactNodeViewRenderer(SkillNodeComponent);
-  },
-
-  addCommands() {
-    return {
-      insertSkillNode:
-        (attrs: SkillNodeAttributes) =>
-        ({ commands }) =>
-          commands.insertContent([
-            {
-              type: SKILL_NODE_TYPE,
-              attrs,
-            },
-            { type: "text", text: " " },
-          ]),
-    };
-  },
-
-  // Markdown serialization and deserialization.
-  markdownTokenizer: {
-    name: SKILL_NODE_TYPE,
-    level: "inline",
-    start: (src) => {
-      const skillTagStart = src.indexOf(`<${SKILL_TAG_NAME}`);
-      const unavailableSkillTagStart = src.indexOf(
-        `<${UNAVAILABLE_SKILL_TAG_NAME}`
-      );
-      const starts = [skillTagStart, unavailableSkillTagStart].filter(
-        (start) => start >= 0
-      );
-
-      return starts.length > 0 ? Math.min(...starts) : -1;
-    },
-    tokenize: (src) => {
-      const match = SKILL_REFERENCE_TAG_REGEX_BEGINNING.exec(src);
-      if (!match) {
-        return undefined;
-      }
-
-      const skill = parseSkillReferenceTag(match[0]);
-      if (!skill) {
-        return undefined;
-      }
-
-      return {
-        type: SKILL_NODE_TYPE,
-        raw: match[0],
-        skillId: skill.id,
-        skillIcon: skill.icon,
-        skillName: skill.name,
-        skillUnavailable: skill.unavailable,
-      };
-    },
-  },
-
-  parseMarkdown: (token) => ({
-    type: SKILL_NODE_TYPE,
-    attrs: {
-      skillId: token.skillId,
-      skillIcon: token.skillIcon,
-      skillName: token.skillName,
-      skillUnavailable: token.skillUnavailable,
-    },
-  }),
-
-  renderMarkdown: (node) => {
-    if (!isString(node.attrs?.skillId)) {
-      return "";
-    }
-
-    if (node.attrs.skillUnavailable === true) {
-      return serializeUnavailableSkillTag({ id: node.attrs.skillId });
-    }
-
-    return isString(node.attrs?.skillName)
-      ? serializeSkillTag({
-          id: node.attrs.skillId,
-          icon: isString(node.attrs.skillIcon) ? node.attrs.skillIcon : null,
-          name: node.attrs.skillName,
-        })
-      : "";
   },
 });
